perf(chat view): make index chunk smaller
